backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
python-multipart==0.0.18
pydantic==2.10.3
python-dotenv==1.0.1
supabase==2.10.0
httpx==0.27.0
